A simple Linux kernel module that prints the kernel version in logs
It's based on materials available in the Bootlin kernel training
materials:
 * https://bootlin.com/doc/training/linux-kernel/linux-kernel-slides.pdf
 * https://github.com/bootlin/training-materials/blob/master/labs/kernel-module-simple/kernel-module-simple.tex

It's not possible to see it from the source code, but the code builds
and works on the BeagleBone Black Wireless. In order to build and
compile it for the BeagleBone you need to run:

export ARCH=arm
export CROSS_COMPILE=arm-linux-gnueabi-
export KDIR=/home/kmu/ws/linux

make

arm-linux-gnueabi- - is a prefix for ARM toolchain used to cross
compile and /home/kmu/ws/linux is the path where I store the kernel code
for the kernel I use on my BeagleBone.