Remove unnecessry delays between I2C messages
Ideally we would not need any delays between the I2C transactions.
However Nintendo Wiichuk has a quirk that we need to issue a read
to cause the device to update its internal state.

Moreover it seems that it actually takes some time for the device
to update its state, so we need to introduce a delay to let the
device update its internal state before reading it.

Before the driver used delays basically before every I2C operation,
but it does not seem necessary and I left just one delay between
two successive reads of the *0x00* register.