Create a simple input device for Nintendo Wiichuk
This change exposes an imput device for the Nintendo Wiichuk. The
device is really simple and only supports key events and two buttons
at this point.

I set the polling interval for the device to 5ms, which seem to be
reasonable given normal human reaction time and should not be
annoyingly slow.