Add support for the Nintendo Wiichuk joystick
Joystick in Linux Kernel is an absolute pointer device. meaning
it reports absolute coordinate events. Nintendo Wiichuk joystick
reports possition in the range [0, 255] in two axis.

The center position in my case for X is 126 and for Y 124. So it
is not perfectly centered, thus I use flat value of 8.