Add a skeleton of the USB driver for the FTDI based device + tools
I have an FTDI produced cable with MPSSE module. Such FTDI devices
are commonly used to convert USB to UART and there is even a default
kernel driver to support them. However MPSSE module actually allows
to convert USB to other protocols like I2C or SPI, but the default
kernel driver does not support this mode of operation.

The cable I have has EEPROM, so it could be used to persist some
configuration on the device itself. So what I want to try to do is
to program the device and give it some fake VID:PID so the kernel
doesn't recognize it as a normal USB-to-UART converter and instead
make an USB-to-I2C convert from it using the MPSSE module
capabilities.

For now I picked VID:PID 0x0005:0x0001 as it appear to be unused
at the moment.

This change includes the skeleton of the USB driver that simply
prints some logs when the device is connect or disconnected. In
addition to the driver skeleton the change includes a few tools to
enumerate the available FTDI devices and override VID and PID on the
device.

For the tools to work they need the FTDI libraries that could be
downloaded from the ftdichip.com. Files ftd2xx.h and WinTypes.h are
part of the FTDI library distribution and therefore they retain the
original FTDI copyright.