Add a library to communicate with MPSSE chip in FTDI
MPSSE is a somewhat generic FTDI chip that can take a set of commands
and translate them into signals on pins of the chip. It can be
programmed to emulate a number of simple protocols like I2C.

The chip presents itself as a two USB pipes (in and out) and to send a
command to MPSSE we need to write data to the out pipe. If the command
results in reading some data by the chip, the data can be collected from
the in pipe. Communicating with MPSSE, once it's configured, is just a
question of reading/writing data over USB.

This change adds a basic library to configure and communicate with MPSSE
and a dummy program to verify that MPSSE was properly configured. Since
the end goal of this is to make MPSSE work as an I2C master I called the
tool i2c_read.

Since i2c_read is not complete it has a bunch of flags that must be
specified, but are not currently used. I'll fix that in the follow up
changes.