Make size parameters of FTDI read/write routines unsigned
It doesn't make much sense for the amount of data to be negative.
The previous behavior converted all the values to unsigned anyway,
so this change just calrifies the interface to make it obvious.

However even if I had cases where negative size values were used
those case are broken anyway already, so this change is hardly
controversial.