Implement basic configuration and data transfer routines for MPSSE
I only picked implemented functions that might be need to implement
I2C over MPSSE, so the implemented function are by no means complete
or generic.

Also note that at this point I didn't really test any of them, so
likely more changes to the implementation will follow up.

Along the way I changed a little bit the structure of the code to
introduce mpsse_io_buffer and mpsse_cmd structures. mpsse_io_buffer
is not a new concept, before this change struct mpsse already
contained a command buffer. This change generalizes this concept
to also manage a buffer for data read from the MPSSE.

mpsse_cmd is just a handler that is used to provide access to the
data read as part of executing a particular command. For each
command we need to know how big the command is (how many bytes we
need to write for the command to MPSSE) and how big the response
is (how many bytes we need to read back from MPSSE as a response
to the command).

At this point the library doesn't have any reasonable device error
handling, but that should be bearable since the end goal is to
have the Linux Kernel driver instead of a userspace tool. That
being said I do intend to reuse the architecutre in the Linux
Kernel driver implementation as well where error handling is not
optional.