Reset FTDI device in the probe
FTDI didn't publish the protocol for the device, so it's a bit of a
magic. Since the documentation is not available, I wrote a program that
using the FTDI userspace driver does the reset and basically nothing
else.

The FTDI userspace driver while doesn't have sources publicly available
uses usbfs. So I enabled usbfs snooping and run the program I created
and this way figured what requests we need to send for the reset.

Note that my reset is slightly different from the reset done in ftdi_sio
module, but there are common parts. That on the one hand makes me think
that there is truth to the reset sequence I figured out, on the other
hand I don't know what the difference between ftdi_sio reset and my
reset means.