Disable special characters and put the device into MPSSE mode
Or at least that's what I think those commands do. The way I figured out
those commands is by checking what the existing FTDI userspace library
does, so it's a bit of magic given that the specification is not
available.

Hopefully this change completes the device reset process. The code only
checks that the requests were successfully submitted. So it's not clear
if the device was actually configured correctly. I suspect that the way
to check it is by reading from the bulk end point because it appears
that every time I do that device always returns 2 bytes. I suspect that
those 2 bytes somehow indicate the device state, but I cannot really use
that since I don't know how to interpret those.

Fortunately, if this code works we don't need to work with control
messages anymore as in MPSSE mode the communication is done completely
via bulk end points and the protocol is documented.