Add basic MPSSE commands and use them to verify MPSSE mode
This change introduces all the essential MPSSE commands required for I2C
communications over FTDI device. I use them to verify that MPSSE mode is
configured.

The way verification works is that we send a couple of incorrect
commands and check the device response. The command is just one byte and
the expected response 0xfa followed by the command byte itself.

The verification is performed as part of the reset procedure for the
device. In addition to that I also initialize MPSSE parameters in order
to use it for I2C communications and put I2C bus into idle mode as it's
expected to be.