Complete and tested I2C implementation for the FTDI MPSSE
This change includes a complete implementation of I2C over FTDI MPSSE
based cable. The implementation has been tested with Nintendo Wiichuck
controller connected to the cable.

The implementation is somewhat different from the userspace
implementation using FTDI library and I find it somewhat more messier.
The reason for that is that FTDI library actually handles internally
some details of the FTDI communication protocol. Since I cannot use the
FTDI library in the kernel I have to deal with it myself.

Aside from that the implementation in the kernel is slightly better at
the error handling. At the very least it tries to reset the device when
it detects IO errors.

Another difference between user space and kernel space implementation
is how I2C Start and Stop conditions are implemented. In the kernel
I use usleep_range to create a delay between changes of of SDA and SCL
values. In user space implementation I just repeated the command
setting the pins multiple times to create a delay.